\name{cityhash.64}
\alias{cityhash.64}
\title{Calculate hash value of input x using CityHash
https://code.google.com/p/cityhash/}
\usage{
  cityhash.64(x)
}
\arguments{
  \item{x}{a number or a character}
}
\value{
  an integer hash value
}
\description{
  Calculate hash value of input x using CityHash
  https://code.google.com/p/cityhash/
}

