\name{murmur3.32}
\alias{murmur3.32}
\title{Calculate hash value of input x using Murmur3
https://code.google.com/p/cityhash/}
\usage{
  murmur3.32(x)
}
\arguments{
  \item{x}{a number or a character}
}
\value{
  an integer hash value
}
\description{
  Calculate hash value of input x using Murmur3
  https://code.google.com/p/cityhash/
}

