\name{spooky.32}
\alias{spooky.32}
\title{Calculate hash value of input x using SpookyHash
http://en.wikipedia.org/wiki/Jenkins_hash_function}
\usage{
  spooky.32(x)
}
\arguments{
  \item{x}{a number or a character}
}
\value{
  an integer hash value
}
\description{
  Calculate hash value of input x using SpookyHash
  http://en.wikipedia.org/wiki/Jenkins_hash_function
}

